add base widget
